refactor: improve userData with useContext
