+++ src/utils/UserContext.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+type User = {
+    id: number
+    name: string
+    description: string
+  }
+
+const usersData = React.createContext<User[]>([]);
+export default usersData
